tasksender/requirements.txt
python-decouple==3.4
pika==1.2.0